+++ react/country/ARG.js
@@ -21239,7 +21239,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
-      required: false,
+      required: true,
       handler: (address) => {
         return {
           ...address,