+++ react/country/AUT.js
@@ -97,6 +97,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
+      required: true,
     },
 
     number: {