+++ react/country/BEL.js
@@ -69,7 +69,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
-      required: false,
+      required: true,
     },
 
     street: {