+++ react/country/CAN.js
@@ -102,7 +102,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
-      required: false,
+      required: true,
     },
 
     number: {