+++ react/country/COL.js
@@ -1288,7 +1288,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
-      required: false,
+      required: true,
       handler: (address) => {
         if (
           !address.state ||