+++ react/country/DEU.js
@@ -104,6 +104,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
+      required: true,
     },
 
     number: {