+++ react/country/ECU.js
@@ -1353,7 +1353,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
-      required: false,
+      required: true,
     },
 
     number: {