+++ react/country/FRA.js
@@ -77,7 +77,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
-      required: false,
+      required: true,
     },
 
     number: {