+++ react/country/HRV.js
@@ -109,6 +109,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
+      required: true,
     },
 
     number: {