+++ react/country/KOR.js
@@ -85,7 +85,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
-      required: false,
+      required: true,
     },
 
     street: {