+++ react/country/PRT.js
@@ -86,6 +86,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
+      required: true,
     },
 
     number: {