+++ react/country/PRY.js
@@ -443,6 +443,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
+      required: false,
     },
 
     complement: {