+++ react/country/SRB.js
@@ -78,6 +78,7 @@ export default {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
+      required: true,
     },
 
     number: {